Move classes from shape subcomponent into spectrum subcomponent; ongoing docs progress
