+++ docs/source/components.rst
@@ -8,14 +8,25 @@ Components
 Components render an entire SWAN command line specification.
 
 * :doc:`components/startup`
+* :doc:`components/cgrid`
+* :doc:`components/inpgrid`
+* :doc:`components/boundary`
 * :doc:`components/physics`
+* :doc:`components/numerics`
+* :doc:`components/output`
+* :doc:`components/lockup`
 
 
 Subcomponents
 -------------
 
-Subcomponents render part of a SWAN command line specification.
+Subcomponents render part of a SWAN command line specification. They typically define
+specific arguments to one or more component.
 
 * :doc:`subcomponents/startup`
+* :doc:`subcomponents/spectrum`
+* :doc:`subcomponents/cgrid`
+* :doc:`subcomponents/time`
+* :doc:`subcomponents/readgrid`
+* :doc:`subcomponents/boundary`
 * :doc:`subcomponents/physics`
-
