+++ tests/swan/subcomponents/test_subcomponent_shape.py
@@ -1,5 +1,5 @@
 """Test SWAN subcomponents."""
-from rompy.swan.subcomponents.shape import JONSWAP, TMA, PM, GAUSS, BIN, SHAPESPEC
+from rompy.swan.subcomponents.spectrum import JONSWAP, TMA, PM, GAUSS, BIN, SHAPESPEC
 
 
 def test_jonswap():
